+++ docs/changelog.md
@@ -18,13 +18,31 @@ Each release rewinds and fast-forwards your mixtape experience—just like a rea
 
 ### 🔧 Changed
 
--
+- On-the-fly Transcoding for playback of flac files
 
 ### 🐛 Fixed
 
 - When searching with multiple words the results narrows down
 
 [v0.3.0](https://github.com/mark-me/mixtape-society/compare/v0.2.2...v0.3.0)
+
+---
+
+## 📦 v0.2.5
+
+<span class="md-tag">Release</span>
+
+**🗓️** 2025-12-29
+
+### 🔧 Changed
+
+- On-the-fly Transcoding for playback of flac files, reducing bandwidth requirements
+
+### 🐛 Fixed
+
+- When searching with multiple words the results narrows down
+
+[v0.3.0](https://github.com/mark-me/mixtape-society/compare/v0.2.2...v0.2.5)
 
 ---